+++ lib/alf/errors.rb
@@ -1,5 +1,6 @@
 module Alf
   class Error < StandardError; end
   class NoSuchRelvarError < Error; end
+  class NoSuchTupleError < Error; end
   class CoercionError < Error; end
 end